% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{hms2sec}
\alias{hms2sec}
\title{Convert HH:MM:SS to seconds}
\usage{
hms2sec(hms_str, anchor = "11:29:00")
}
\arguments{
\item{anchor}{the initial time point to calculate relative lag seconds}

\item{hmm_str}{the time string in form as "11:29:10"}
}
\value{
a value or vector of relative lag seconds
}
\description{
Convert strings HH:MM:SS into relative seconds distance from 11:29:00,
as the final minute puts more on the possibility of winning a bid.
}

