% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_months.R
\name{trend_months}
\alias{trend_months}
\title{Visualize price tendencies in months}
\usage{
trend_months(years = 2016:2016, months = 1:6, min_lag = 0, max_lag = 60)
}
\arguments{
\item{years}{a vector of years to plot}

\item{months}{a vector of months to plot}

\item{min_lag}{the minimum time lag comparing to 11:29:00, which is for reading briefness.}

\item{max_lag}{the maximum time lag comparing to 11:29:00.}
}
\value{
a gglot2 instance
}
\description{
The delta price (real - warnning) of each month is plot, which helps to
compare the global tendency of price increment during last one or two minutes.
}

